requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
motor>=3.3.0
pymongo[srv]>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
minio>=7.2.0
python-dotenv>=1.0.0
email-validator>=2.1.0
httpx>=0.26.0
certifi>=2024.2.2
dnspython>=2.5.0
